Testing RQ implementation
