+++ run.py
@@ -1,3 +1,3 @@
-from app import app
+from app2 import app
 
 app.run()
